% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APF.R
\name{APF}
\alias{APF}
\title{Log-Sum-Exp function to prevent underflow}
\usage{
APF(psi_pa, data, N, model)
}
\arguments{
\item{psi_pa}{psi parameters}

\item{data}{data that include observations}

\item{N}{number of particle needed}

\item{model}{model used}
}
\value{
A list containing:
}
\description{
Log-Sum-Exp function to prevent underflow
}
