% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_psi.R
\name{combine_psi}
\alias{combine_psi}
\title{Function generating preferred psi function arguments}
\usage{
combine_psi(psi, index)
}
\arguments{
\item{psi}{Parameters of the psi functions}

\item{index}{Indicator of which psi is preferred at each time}
}
\value{
Preferred psi function arguments
}
\description{
This function takes the parameters of the psi functions produced by each 
of the two sets of iAPF and combines them to give a single preferred value at each
time. Easily extensible to more than 2 iAPFs if needed.
}
