% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dKS.R
\name{compute_dKS}
\alias{compute_dKS}
\title{Function to compute mahalanobis distance between weighted mean and fks mean}
\usage{
compute_dKS(x, w, fks.obj)
}
\arguments{
\item{x}{Particles generated by the algorithm}

\item{w}{Weights of particles}

\item{fks.obj}{Fks information of the linear Gaussian model}
}
\value{
A plot contains the distance of the whole time scale
}
\description{
This function aims to evaluate whether the estimated smoothing distribution is close to the real smoothing
distribution computed by FKS for Gaussian linear models by evaluating the distance between them
}
