% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_twisted_potential.R
\name{eval_twisted_potential}
\alias{eval_twisted_potential}
\title{Function to provide the log-density of a twisted potential function}
\usage{
eval_twisted_potential(model, psi_pa, x, log_likelihoods)
}
\arguments{
\item{model}{List containing model parameters}

\item{psi_pa}{Parameters of psi function}

\item{x}{States at which to evaluate}

\item{log_likelihoods}{log_likelihoods of the data point at which to evaluate}
}
\value{
Twisted potential function evaluated at x
}
\description{
This function evaluates a twisted likelihood when provided with the likelihood function, the
twisting function, the observation and the locations at which to perform the evaluation.
}
