% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_log_g.R
\name{evaluate_log_g}
\alias{evaluate_log_g}
\title{Function to evaluate the log-density of the observation density}
\usage{
evaluate_log_g(model, x, datum)
}
\arguments{
\item{model}{List containing model parameters}

\item{x}{State at which to evaluate}

\item{datum}{Data point at which to evaluate}
}
\value{
Log-density of the observation density
}
\description{
This function evaluates the log-density of the observation density. 
It is expected to be updated to evaluate log-density of general potential functions.
}
