% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_blocks_half.R
\name{generate_blocks_half}
\alias{generate_blocks_half}
\title{Function to generate blocks that the iAPF runs within}
\usage{
generate_blocks_half(lag, len)
}
\arguments{
\item{lag}{Lag of the blocks user specified}

\item{len}{Length of the time scale}
}
\value{
A list with two elements:
breaks is the break points on the time scale that the iAPF starts to iterate
psi_index indicates at each time point, the outputs generated by which iAFP layer the algorithm chooses
}
\description{
Function to generate blocks that the iAPF runs within
}
