% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iAPF.R
\name{iAPF}
\alias{iAPF}
\title{Log-Sum-Exp function to prevent underflow}
\usage{
iAPF(data, Napf, model)
}
\arguments{
\item{data}{data used}

\item{Napf}{number of particles}

\item{model}{model used}
}
\value{
A list containing:
}
\description{
Log-Sum-Exp function to prevent underflow
}
