% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_psi.R
\name{learn_psi}
\alias{learn_psi}
\title{Function to learn twisting psi function parameters}
\usage{
learn_psi(x, obs, model)
}
\arguments{
\item{x}{A collection of particle locations}

\item{obs}{Observations}

\item{model}{List containing model parameters}
}
\value{
Twisting psi function parameters
}
\description{
This function takes a collection of particle locations and calculates twisting psi function arguments.
}
