% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num.R
\name{Num}
\alias{Num}
\title{Rolling window function}
\usage{
Num(Z, l, k)
}
\arguments{
\item{Z}{normalizing constant estimate}

\item{l}{number of iteration}

\item{k}{user-specified constant}
}
\value{
Number of particles next iteration needed
}
\description{
Rolling window function
}
