% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orc_smc.R
\name{Orc_SMC}
\alias{Orc_SMC}
\title{Rolling window function}
\usage{
Orc_SMC(lag, data, model, N)
}
\arguments{
\item{lag}{}

\item{data}{}

\item{model}{}

\item{N}{}
}
\value{
A matrix of observations with dimensions Time x d.
}
\description{
Rolling window function
}
