% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_particles.R
\name{resample_particles}
\alias{resample_particles}
\title{Function to resample}
\usage{
resample_particles(log_weights)
}
\arguments{
\item{log_weights}{log weights}
}
\value{
A list containing:
}
\description{
Function to resample
}
