% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpf.R
\name{run_bpf}
\alias{run_bpf}
\title{Rolling window function}
\usage{
run_bpf(model, data, lag, Napf)
}
\arguments{
\item{model}{model used}

\item{data}{data used}

\item{lag}{lag}

\item{Napf}{Number of particles}
}
\value{
A list containing
}
\description{
Rolling window function
}
