% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iAPF.R
\name{run_iAPF}
\alias{run_iAPF}
\title{Function to run the iterated APF}
\usage{
run_iAPF(model, data, Napf)
}
\arguments{
\item{model}{Information of the model}

\item{data}{Information of the required data}

\item{Napf}{Number of particles users would like to generate}
}
\value{
A list contains:
X_apf is the particles generated by iAPF at the time of the upper boundary of the given block
w_apf is the weights generated by iAPF at the time of the upper boundary of the given block
psi_pa is the parameters of the distribution psi the algorithm finally choose
Z_apf is the estimated normalising constant
ancestors is all the indices of ancestors of particles during the algorithm
}
\description{
Function to run the iterated APF
}
