% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_psi_APF.R
\name{run_psi_APF}
\alias{run_psi_APF}
\title{Function of psi_APF}
\usage{
run_psi_APF(model, data, N, psi_pa, init)
}
\arguments{
\item{model}{List containing model parameters}

\item{data}{List containing time series on which to run the filter}

\item{N}{Number of particles to use}

\item{psi_pa}{Parameters of the twisting function}

\item{init}{}
}
\value{
A list containing:
x is the particle set generated during the sampling and resampling approach
w is the weights of particles
ancestors is indices associated with every resampling event
logZ is normalizing constant estimate
resample_time is the specific time that resample happens
}
\description{
This function performs sampling and resampling procedures of
psi-Auxiliary Particle Filter with κ-adaptive resampling
}
