% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observation_lg.R
\name{simulate_observation}
\alias{simulate_observation}
\title{Function to simulate observations}
\usage{
simulate_observation(state, params)
}
\arguments{
\item{state}{State which to evaluate observations at}

\item{params}{Parameters of the observation distribution}
}
\value{
Observations generated at the specific state
}
\description{
Function to simulate observations
}
