% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observation_msv.R
\name{simulate_observation_msv}
\alias{simulate_observation_msv}
\title{Function to simulate observations for multivariate svm model}
\usage{
simulate_observation_msv(x, obs_params)
}
\arguments{
\item{state}{State which to evaluate observations at}

\item{params}{Parameters of the observation distribution}
}
\value{
Observations generated at the specific state
}
\description{
Function to simulate observations for multivariate svm model
}
