% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_psi.R
\name{learn_psi}
\alias{learn_psi}
\title{Function to learn twisting log_psi function parameters}
\usage{
learn_psi(x, obs, model, log_likelihoods)
}
\arguments{
\item{x}{A collection of particle locations}

\item{obs}{Observations}

\item{model}{List containing model parameters}

\item{log_likelihoods}{log_likelihoods of the particles generated during the iAPF}
}
\value{
Twisting log_psi function parameters
}
\description{
This function takes a collection of particle locations and calculates twisting log_psi function arguments.
}
