% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iAPF.R
\name{run_iAPF}
\alias{run_iAPF}
\title{Function to run two layers of the iterated APF}
\usage{
run_iAPF(model, data, Napf)
}
\arguments{
\item{model}{Information of the model}

\item{data}{Observations}

\item{Napf}{Number of particles users would like to generate}
}
\value{
A list contains:
X_apf is all the particles generated by iAPF
w_apf is the weights of particles
psi_final is the parameters of the distribution psi the algorithm finally choose
Z_apf is the estimated normalising constant
ancestors is all the indices of ancestors of particles during procedures
}
\description{
Function to run two layers of the iterated APF
}
